+++ src/command_modules/azure-cli-monitor/HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 
 0.0.12
 ++++++
+* Add activity-log alert commands
 * Minor fixes.
 
 0.0.11